% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_quotes}
\alias{add_quotes}
\title{Put vector values in quotes}
\usage{
add_quotes(x, quote = "double")
}
\arguments{
\item{x}{vector of string / numeric}

\item{quote}{what type of quotes (\code{double} or \code{single})}
}
\value{
Character vector of input with quotation marks around each value
}
\description{
Put vector values in quotes
}
