% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ruv.R
\name{add_ruv}
\alias{add_ruv}
\title{Add residual variability to the dependent variable}
\usage{
add_ruv(x, ruv = list())
}
\arguments{
\item{x}{dependent value without residual variability}

\item{ruv}{list specifying proportional, additive and/or exponential errors (`prop`, `add`, `exp`)}
}
\description{
Add residual variability to the dependent variable
}

