% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence.R
\name{adherence_binomial}
\alias{adherence_binomial}
\title{Binomial adherence}
\usage{
adherence_binomial(n = 100, prob)
}
\arguments{
\item{n}{number of occasions}

\item{prob}{binomial probability}
}
\value{
Returns a vector of length `n`
  containing values 0 (non-adherent) or 1 (adherent).

Numeric vector of length n
}
\description{
Model adherence as a binomial probability at the time of each occasion.
}
