% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{advan}
\alias{advan}
\title{ADVAN-style functions to calculate linear PK systems}
\usage{
advan(model, cpp = TRUE)
}
\arguments{
\item{model}{Standard linear PK model, e.g. \verb{1cmt_iv_bolus}.}

\item{cpp}{use C++-versions of model (~50x faster than R implementations)}
}
\value{
Model function
}
\description{
ADVAN-style functions to calculate linear PK systems
}
