% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan_create_data.R
\name{advan_create_data}
\alias{advan_create_data}
\title{Create ADVAN-style dataset}
\usage{
advan_create_data(
  regimen,
  parameters,
  cmts = 5,
  t_obs = NULL,
  covariates = NULL,
  covariate_model = NULL
)
}
\arguments{
\item{regimen}{PKPDsim regimen}

\item{parameters}{list of parameters}

\item{cmts}{number of compartments, minimum is 1. Default is 5, which is enough for most linear PK models. It is OK to have more compartments available than are actually being used.}

\item{t_obs}{add observation timepoints to dataset}

\item{covariates}{covariate list}

\item{covariate_model}{covariate model equations, written in C}
}
\value{
Data frame of ADVAN-style data
}
\description{
Create ADVAN-style dataset
}
