% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\docType{data}
\name{advan_funcs}
\alias{advan_funcs}
\title{ADVAN-style functions to calculate linear PK systems}
\format{An object of class \code{list} of length 9.}
\usage{
advan_funcs
}
\description{
ADVAN-style functions to calculate linear PK systems
}
\keyword{datasets}
