% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan_parse_output.R
\name{advan_parse_output}
\alias{advan_parse_output}
\title{Internal function to parse the raw output from ADVAN-style functions}
\usage{
advan_parse_output(data, cmts = 1)
}
\arguments{
\item{data}{simulation output data}
}
\description{
Internal function to parse the raw output from ADVAN-style functions
}
