% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan_process_infusion_doses.R
\name{advan_process_infusion_doses}
\alias{advan_process_infusion_doses}
\title{Add column RATEALL to ADVAN-style dataset to handle infusions}
\usage{
advan_process_infusion_doses(data)
}
\arguments{
\item{data}{ADVAN-style dataset, e.g. created using `advan_create_data`.}
}
\description{
Add column RATEALL to ADVAN-style dataset to handle infusions
}
