% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_eqn_wrapper.R
\name{analytical_eqn_wrapper}
\alias{analytical_eqn_wrapper}
\title{Wrapper for using analytical equations with PKPD regimens}
\usage{
analytical_eqn_wrapper(analytical, design = NULL, parameters)
}
\arguments{
\item{analytical}{analytical equation, taking parameters \code{amt}, \code{parameters}, and \code{t}, and returning a vector of values for \code{y}}

\item{design}{design dataset created by \code{sim_ode}}

\item{parameters}{list of parameters}
}
\description{
In development. Needs to be optimized significantly to be useful in production.
}
\keyword{internal}
