% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lagtime.R
\name{apply_lagtime}
\alias{apply_lagtime}
\title{Apply lagtime to a regimen}
\usage{
apply_lagtime(regimen, lagtime, parameters)
}
\arguments{
\item{regimen}{PKPDsim regimen}

\item{lagtime}{lagtime object, either single value / parameter name or vector of values/parameter names for all compartments.}

\item{parameters}{parameter list, required if parameters are specified.}
}
\description{
Apply lagtime to a regimen
}
