% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_default_literature_model.R
\name{available_default_literature_models}
\alias{available_default_literature_models}
\title{See models from the literature available for installation}
\usage{
available_default_literature_models()
}
\value{
Returns a character vector of models available for installation
}
\description{
See models from the literature available for installation
}
\examples{
available_default_literature_models()
}
