% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_to_R_code.R
\name{bioavailability_to_R_code}
\alias{bioavailability_to_R_code}
\title{Transforms bioavailability specs into appropriate R code}
\usage{
bioavailability_to_R_code(bioav)
}
\arguments{
\item{bioav}{bioavailability specification, either NULL (assume a value of 1
in all compartments), a single value (assume it applies to all
compartments), or a vector of values.}
}
\value{
character string of length 1
}
\description{
Specialized wrapper around `vector_to_R_code` that makes reasonable PK
assumptions for when the bioavailability specification is NULL.
}
