% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_y.R
\name{calc_dydP}
\alias{calc_dydP}
\title{Calculate derivative}
\usage{
calc_dydP(dy, y, rel_delta, log_y)
}
\arguments{
\item{dy}{dy}

\item{y}{dependent value}

\item{rel_delta}{relative delta}

\item{log_y}{logical indicating if the dependent variable is log transformed}
}
\description{
Calculate derivative
}
