% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_lin.R
\name{calc_ss_lin}
\alias{calc_ss_lin}
\title{Returns the state of a linear PK system at steady state (trough)}
\usage{
calc_ss_lin(
  f = "TwoCompOral",
  dose,
  interval,
  model,
  parameters,
  covariates = NULL,
  map = NULL,
  n_days = 45,
  n_transit_compartments = 0,
  auc = FALSE
)
}
\arguments{
\item{dose}{dose}

\item{interval}{interval}

\item{model}{PKPDsim model}

\item{parameters}{parameters list}

\item{covariates}{covariates list}

\item{map}{list for remapping parameters}

\item{n_days}{number of days at which to assume steady state. Default is 45.}

\item{n_transit_compartments}{number of transit compartments, will insert n compartments between the first (dose) compartment and the second (central) compartment.}

\item{auc}{add (empty) AUC compartment at end of state vector?}
}
\description{
Basically it performs a PK simulation using algebraic equations instead
of ODEs to steady state (n=45 days, increased if needed). This is much
faster than doing the same using ODEs.
}
\details{
The function needs to be made a bit more generic, it focused now
on the TwoCompOral model.

It can also be used for models with transit compartments, however, 
the assumption is made that at the end of the dosing interval the 
amount in the transit compartments is negligible (0).
}
