% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe/chain an object forward into a function call/expression.}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{a value}

\item{rhs}{a function/call/expression. Enclose in parentheses to force
evaluation of rhs before piping is carried out, e.g. anonymous functions
or call-generating expressions.}
}
\value{
The result of evaluting the right-hand side with the left-hand side
as the specified argument(s).
}
\description{
Pipe/chain an object forward into a function call/expression.
}
