% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_iov_specification.R
\name{check_iov_specification}
\alias{check_iov_specification}
\title{Checks that IOV was specified appropriately}
\usage{
check_iov_specification(iov, code, pk_code)
}
\arguments{
\item{iov}{IOV specifications, provided as a nested named list.}

\item{code}{C++ ODE code, supplied as a string}

\item{pk_code}{C++ PK code, supplied as a string}
}
\description{
Inter-occasion variability (IOV) is expected to be supplied as a list with
`cv` and `n_bins` specified. `cv` is expected to be a named list with IOV
for each PK parameter. This function then checks to ensure that the PK code
or ODE code contains an IOV term for each PK parameter specified.
}
\keyword{internal}
