% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mixture_model.R
\name{check_mixture_model}
\alias{check_mixture_model}
\title{Check that mixture model is specified in right format and within constraints (1 parameter, 2 groups)}
\usage{
check_mixture_model(mixture, parameters)
}
\arguments{
\item{mixture}{mixture model specification (as list, e.g. `list("CL" = list(values=c(5, 10), probability=0.3))`)}

\item{parameters}{vector of parameter names}
}
\description{
Check that mixture model is specified in right format and within constraints (1 parameter, 2 groups)
}
\keyword{internal}
