% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_obs_types.R
\name{check_obs_input}
\alias{check_obs_input}
\title{Checks obs input for valid combinations of cmt, var, scale}
\usage{
check_obs_input(obs)
}
\arguments{
\item{obs}{specified observation object including at least a description of
which variable(s) are associated with a particular compartment, e.g.
`list(variable="CONC", scale="1")`.}
}
\description{
Checks obs input for valid combinations of cmt, var, scale
}
