% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_sim_cpp.R
\name{compile_sim_cpp}
\alias{compile_sim_cpp}
\title{Compile ODE model to c++ function}
\usage{
compile_sim_cpp(code, size, p, cpp_show_code, code_init = NULL,
  declare_variables = NULL, covariates = NULL, obs = NULL, dose = NULL,
  verbose = FALSE)
}
\arguments{
\item{code}{C++ code}

\item{size}{size of ODE system}

\item{p}{parameters (list)}

\item{cpp_show_code}{show output c++ function?}

\item{code_init}{code for initialization of state}

\item{declare_variables}{variable declaration}

\item{covariates}{covariates specification}

\item{obs}{observation specification}

\item{dose}{dose specification}

\item{verbose}{show more output}
}
\description{
Compile ODE model to c++ function
}

