% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_last_obs_only.R
\name{covariate_last_obs_only}
\alias{covariate_last_obs_only}
\title{Use only last observed covariate values}
\usage{
covariate_last_obs_only(covariates)
}
\arguments{
\item{covariates}{covariates object}
}
\description{
Use only last observed covariate values
}
