% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_table_to_list.R
\name{covariate_table_to_list}
\alias{covariate_table_to_list}
\title{Convert covariate table specified as data.frame
Can handle time-varying data too, if `t` or `time` is specified as column}
\usage{
covariate_table_to_list(covariates_table, covariates_implementation = list())
}
\arguments{
\item{covariates_table}{`data.frame`` with covariates in columns. Potentially with `id` and `t` columns}

\item{covariates_implementation}{`list` with implementation method per covariate}
}
\description{
Convert covariate table specified as data.frame
Can handle time-varying data too, if `t` or `time` is specified as column
}

