% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates_table_to_list.R
\name{covariates_table_to_list}
\alias{covariates_table_to_list}
\title{Convert covariate table specified as data.frame}
\usage{
covariates_table_to_list(covariates_table, covariates_implementation = list())
}
\arguments{
\item{covariates_table}{\verb{data.frame`` with covariates in columns. Potentially with }id\code{and}t` columns}

\item{covariates_implementation}{\code{list} with implementation method per covariate}
}
\value{
List of covariates
}
\description{
Can handle time-varying data too, if \code{t} or \code{time} is specified as column
}
