% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_table.R
\name{create_event_table}
\alias{create_event_table}
\title{Create an event table}
\usage{
create_event_table(
  regimen,
  t_max = NULL,
  t_obs = NULL,
  t_tte = NULL,
  t_init = 0,
  p,
  covariates,
  model = NULL,
  obs_type = NULL
)
}
\arguments{
\item{regimen}{regimen}

\item{t_max}{t_max}

\item{t_obs}{t_obs}

\item{t_tte}{t_tte}

\item{t_init}{t_init}

\item{p}{parameters}

\item{covariates}{covariates}

\item{model}{model}

\item{obs_type}{observation type}
}
\description{
Create an event table
}
