% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs_data.R
\name{create_obs_data}
\alias{create_obs_data}
\title{Create obs data}
\usage{
create_obs_data(ode_data, obs_attr, id)
}
\arguments{
\item{ode_data}{data frame of output from ode() function}

\item{obs_attr}{"obs" attribute from ode() function}

\item{id}{ID of the individual}
}
\description{
Used by \code{\link[=sim]{sim()}} to arrange data from ode() function into the correct format.
}
\seealso{
\code{\link[=sim]{sim()}}
}
\keyword{internal}
