% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_tdm_init_model.R
\name{define_tdm_init_model}
\alias{define_tdm_init_model}
\title{defines C code for TDM before dose conditions}
\usage{
define_tdm_init_model(def)
}
\arguments{
\item{def}{model definition, named recursive list with at least the
objects `misc$model_type`, `parameters` and `variables`}
}
\value{
model defintion with `state_init` object added describing how to initializing the compartments.
}
\description{
Currently only available for 1-cmt and 2-cmt IV models
}
