% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ode_syntax.R
\name{detect_ode_syntax}
\alias{detect_ode_syntax}
\title{Auto-detect the syntax for the ODE code}
\usage{
detect_ode_syntax(code)
}
\arguments{
\item{code}{character string with ODE code}
}
\value{
List with elements \code{from} and \code{to} indicating the syntax for the ODE
code
}
\description{
Either PKPDsim or RxODE
}
