% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ode_model_size.R
\name{get_ode_model_size}
\alias{get_ode_model_size}
\title{Get the number of states in the ODE from the code
code C++ code for model}
\usage{
get_ode_model_size(code)
}
\arguments{
\item{code}{C++ code}
}
\value{
Number of states in the ODE model
}
\description{
Get the number of states in the ODE from the code
code C++ code for model
}
