% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_from_code.R
\name{get_parameters_from_code}
\alias{get_parameters_from_code}
\title{Get model parameters from code}
\usage{
get_parameters_from_code(code, state_init, declare_variables = NULL)
}
\arguments{
\item{code}{code}

\item{state_init}{state init vector}

\item{declare_variables}{declared variables}
}
\value{
Vector of parameter names
}
\description{
Get model parameters from code
}
