% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_t_obs_from_regimen.R
\name{get_t_obs_from_regimen}
\alias{get_t_obs_from_regimen}
\title{Extract sensible default observation times from a specified regimen}
\usage{
get_t_obs_from_regimen(regimen, obs_step_size, t_max, covariates, extra_t_obs)
}
\description{
Extract sensible default observation times from a specified regimen
}

