% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkpd_hemtox.R
\name{hemtox}
\alias{hemtox}
\title{ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)}
\usage{
hemtox(t, A, p)
}
\description{
ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)
}

