% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifelse0.R
\name{ifelse0}
\alias{ifelse0}
\title{ifelse function but then based on whether value is NULL or not}
\usage{
ifelse0(value = NULL, alternative = NULL, allow_null = FALSE)
}
\arguments{
\item{value}{metadata list object}

\item{alternative}{alternative value}

\item{allow_null}{can the alternative be NULL?}
}
\value{
\code{value} if non-NULL; \code{alternative} otherwise
}
\description{
ifelse function but then based on whether value is NULL or not
}
