% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_default_literature_model.R
\name{install_default_literature_model}
\alias{install_default_literature_model}
\title{Install default literature model}
\usage{
install_default_literature_model(model, ...)
}
\arguments{
\item{model}{Name of model, e.g., "pk_busulfan_mccune". See
\code{\link[=available_default_literature_models]{available_default_literature_models()}}}

\item{...}{arguments passed onto \code{model_from_api}. For fine-grain control, it
is better to install models directly from \code{\link[=model_from_api]{model_from_api()}} or
\code{\link[=new_ode_model]{new_ode_model()}}.}
}
\description{
A very lightweight wrapper for \code{model_from_api} that installs previously
published models packaged within PKPDsim.
}
\examples{
\dontrun{
install_default_literature_model("pk_busulfan_mccune")
}
}
