% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_regimen.R
\name{is_bolus_or_oral}
\alias{is_bolus_or_oral}
\title{Is regimen type oral or bolus? (i.e., treat t_inf as zero)}
\usage{
is_bolus_or_oral(regimen)
}
\arguments{
\item{regimen}{object of class "regimen"`}
}
\value{
Returns a logical vector of length equal to the number of doses in
the regimen
}
\description{
Any regimen matching exactly the string 'bolus' or containing the string
'oral' will return \code{TRUE}, otherwise \code{FALSE}. Bolus/oral regimens are
generally treated as having an infusion duration of zero.
}
