% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_newer_package.R
\name{is_newer_package}
\alias{is_newer_package}
\title{Check if package number is different from currently installed,
and provide some messaging.}
\usage{
is_newer_package(package, new_version)
}
\arguments{
\item{package}{R package}

\item{new_version}{new version number}
}
\description{
Technically it only checks if a package version is different, not necessarily
a higher version number.
}
\keyword{internal}
