% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_cov_and_par.R
\name{join_cov_and_par}
\alias{join_cov_and_par}
\title{Combines covariates and parameters into a single list, useful for reparametrization of the model.}
\usage{
join_cov_and_par(covs, pars)
}
\arguments{
\item{covs}{covariates object}

\item{pars}{model parameters, such as the output of the `parameters()` call frmo a model library.}
}
\value{
List containing covariates and parameters
}
\description{
Combines covariates and parameters into a single list, useful for reparametrization of the model.
}
