% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_regimen.R
\name{join_regimen}
\alias{join_regimen}
\title{Join two dosing regimens}
\usage{
join_regimen(regimen1 = NULL, regimen2 = NULL, interval = NULL,
  dose_update = NULL)
}
\arguments{
\item{regimen1}{first regimen}

\item{regimen2}{second regimen}

\item{interval}{interval between regimen1 and regimen2 (if dose_update not specified)}

\item{dose_update}{dose number at which to override regimen1 with regimen 2 (if interval not specified)}
}
\description{
Join two dosing regimens
}

