% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lifespan_idr.R
\name{lifespan_idr}
\alias{lifespan_idr}
\title{DDE system for basic lifespan IDR model}
\usage{
lifespan_idr(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
DDE system for basic lifespan IDR model
}

