% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_to_full.R
\name{lower_triangle_mat_size}
\alias{lower_triangle_mat_size}
\title{Size of the lower triangle of the matrix}
\usage{
lower_triangle_mat_size(mat)
}
\arguments{
\item{mat}{omega matrix as a vector}
}
\description{
Size of the lower triangle of the matrix
}
