% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_regimen.R
\name{merge_regimen}
\alias{merge_regimen}
\title{Merge two regimens together.}
\usage{
merge_regimen(regimens)
}
\arguments{
\item{regimens}{List of PKPDsim regimens created with \code{new_regimen}.}
}
\value{
Merged regimens
}
\description{
In contrast to \code{join_regimen}, which joins two consecutive regimens together, \code{merge_regimen} merges two or
more regimens given at the same time. This can e.g. be used to define regimens for multi-drug models.
}
