% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_from_api.R
\name{model_from_api}
\alias{model_from_api}
\title{Load model definition from API, and compile to R library}
\usage{
model_from_api(model = NULL, url = NULL, verbose = TRUE, ...)
}
\arguments{
\item{model}{model id}

\item{url}{URL of API}
}
\description{
Load model definition from API, and compile to R library
}

