% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_library.R
\name{model_library}
\alias{model_library}
\title{Model library}
\usage{
model_library(name = NULL)
}
\arguments{
\item{name}{name of model in library. If none specified, will show list of available models.}
}
\description{
Model library
}

