% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_library.R
\name{model_to_library}
\alias{model_to_library}
\title{Save a model as an R library}
\usage{
model_to_library()
}
\arguments{
\item{model}{PKPDsim}

\item{folder}{folder for new library}

\item{parameters}{default parameter list}
}
\description{
Save the PKPDsim model as a compiled model. Benefit is much faster load
times, e.g. for use in Shiny app.
}

