% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm2.R
\name{mvrnorm2}
\alias{mvrnorm2}
\title{More powerful multivariate normal sampling function}
\usage{
mvrnorm2(n, mu, Sigma, exponential = FALSE, sequence = NULL, ...)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean}

\item{Sigma}{covariance matrix}

\item{exponential}{exponential distribution (i.e. multiply mu by exponential of sampled numbers)}

\item{sequence}{any sequence available in the randtoolbox, e.g. `halton`, or `sobol`}

\item{...}{parameters passed to mvrnorm or randtoolbox sequence generator}
}
\value{
Multivariate normal samples
}
\description{
Besides standard multivariate normal sampling (mvrnorm), allows exponential
multivarate normal and quasi-random multivariate normal (using the randtoolbox)
all using the same interface.
}
