% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_locf.R
\name{na_locf}
\alias{na_locf}
\title{Adapted from zoo::na.locf0}
\usage{
na_locf(object, fromLast = FALSE)
}
\arguments{
\item{object}{an object}

\item{fromLast}{logical. Causes observations to be carried backward rather than forward. Default is FALSE. With a value of TRUE this corresponds to NOCB (next observation carried backward). It is not supported if x or xout is specified.}
}
\description{
Adapted from zoo::na.locf0
}
