% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adherence.R
\name{new_adherence}
\alias{new_adherence}
\title{Adherence function}
\usage{
new_adherence(n = 100, type = "markov", markov = list(p11 = 0.75, p01 =
  0.75), p_binom = 0.7)
}
\description{
Adherence function
}

