% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_covariate_model.R
\name{new_covariate_model}
\alias{new_covariate_model}
\title{covariate model function}
\usage{
new_covariate_model(model = list())
}
\arguments{
\item{model}{covariate model specified as list}
}
\value{
List containing model function(s)
}
\description{
covariate model function
}
