% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_ode_model.R
\name{new_ode_model}
\alias{new_ode_model}
\title{Create new ODE model}
\usage{
new_ode_model(model = NULL, code = NULL, file = NULL, func = NULL,
  state_init = NULL, parameters = NULL, size = NULL, obs = list(cmt = 1,
  scale = 1), dose = list(cmt = 1), covariates = NULL,
  declare_variables = NULL, cpp_show_code = FALSE, verbose = FALSE)
}
\arguments{
\item{obs}{list with "scale": character string with definition for scale, e.g. "V" or "V*(WT/70)". If NULL, scale defaults to 1., and "cmt" the observation compartment}
}
\description{
Create new ODE model
}

