% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/num_int_wrapper.R
\name{num_int_wrapper}
\alias{num_int_wrapper}
\title{Wrapper for deSolve}
\usage{
num_int_wrapper(times, A_init, des, p_ind, lsoda_func)
}
\arguments{
\item{times}{time vector}

\item{A_init}{vector of initial state}

\item{des}{ODE function}

\item{p_ind}{parameters}

\item{lsoda_func}{LSODA function}
}
\description{
Wrapper for deSolve
}

