% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{OneCompIVbolus}
\alias{OneCompIVbolus}
\title{ADVAN-style equations}
\usage{
OneCompIVbolus(d)
}
\arguments{
\item{d}{data}
}
\description{
Adapted from Abuhelwa et al. JPET 2015
}
\details{
Functions for calculating drug amount in each compartments of the
common pharmacokinetic models (1,2,3 compartment IV bolus, IV infusion,
and first-order absorption models)

 Definitions:
- A*last: is the initial amount at the beginning of each time interval (t, t=t2-t1)
  of a corresponding compartment (i.e. drug amount at the end of the last time interval)
- E* : the sum of Exit (elimination) rate constant of the corresponding compartment.
IV bolus- 1 compartment
}
