% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lagtime.R
\name{parse_lagtime}
\alias{parse_lagtime}
\title{Parse lagtime specified to main sim() function}
\usage{
parse_lagtime(lagtime, ode, parameters)
}
\arguments{
\item{lagtime}{either a value (numeric) or a parameter (character) or NULL.}

\item{ode}{function describing the ODE system}

\item{parameters}{model parameters}
}
\value{
a vector of character or numeric values
}
\description{
Parse lagtime specified to main sim() function
}
