% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_obs_types.R
\name{parse_obs_types}
\alias{parse_obs_types}
\title{Parse observation types to simulation code}
\usage{
parse_obs_types(obs)
}
\arguments{
\item{obs}{specified observation object, e.g. `list(variable="CONC", scale="1")` or `list(cmt=1, scale="V/1000.0")`.}
}
\description{
Parse observation types to simulation code
}
