% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_1cmt_iv.R
\name{pk_1cmt_iv}
\alias{pk_1cmt_iv}
\title{ODE system for PK - 1 compartment IV}
\usage{
pk_1cmt_iv(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PK - 1 compartment IV
}

