% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_1cmt_oral.R
\name{pk_1cmt_oral}
\alias{pk_1cmt_oral}
\title{ODE system for PK - 1 compartment oral administration}
\usage{
pk_1cmt_oral(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PK - 1 compartment oral administration
}

