% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pk_1cmt_oral_pd_viral}
\alias{pk_1cmt_oral_pd_viral}
\title{ODE system for PKPD - 1 compartment IV, HIV viral dynamics model (Perelson)}
\usage{
pk_1cmt_oral_pd_viral(t, A, p)
}
\description{
ODE system for PKPD - 1 compartment IV, HIV viral dynamics model (Perelson)
}

