% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_2cmt_iv_mm.R
\name{pk_2cmt_iv_mm}
\alias{pk_2cmt_iv_mm}
\title{ODE system for PK - 2 compartment IV Michaelis-Menten (non-linear)}
\usage{
pk_2cmt_iv_mm(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PK - 2 compartment IV Michaelis-Menten (non-linear)
}

