% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pk_tte_1cmt_oral_exp_hazard}
\alias{pk_tte_1cmt_oral_exp_hazard}
\title{ODE system for PK-TTE - 1 compartment oral + hazard}
\usage{
pk_tte_1cmt_oral_exp_hazard(t, A, p)
}
\description{
ODE system for PK-TTE - 1 compartment oral + hazard
}

