% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkdata.R
\docType{data}
\name{pkdata}
\alias{pkdata}
\title{PK dataset}
\format{A data frame with 624 rows and 12 variables in NONMEM format}
\usage{
data(pkdata)
}
\description{
Example PK dataset
}
\keyword{datasets}

