% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pkpd_hemtox}
\alias{pkpd_hemtox}
\title{ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)}
\usage{
pkpd_hemtox(t, A, p)
}
\description{
ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)
}

