% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.covariate.R
\name{print.covariate}
\alias{print.covariate}
\title{Print function for PKPDsim covariate object}
\usage{
\method{print}{covariate}(x, ...)
}
\arguments{
\item{x}{covariate object}

\item{...}{additional arguments}
}
\description{
Print function for PKPDsim covariate object
}
