% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PKPDsim.R
\name{print.PKPDsim}
\alias{print.PKPDsim}
\title{Print function for PKPDsim simulation function}
\usage{
\method{print}{PKPDsim}(x, ...)
}
\arguments{
\item{x}{function}

\item{...}{additional arguments}
}
\value{
No return value, print function.
}
\description{
Print function for PKPDsim simulation function
}
