% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.regimen.R
\name{print.regimen}
\alias{print.regimen}
\title{Print function for PKPDsim regimen}
\usage{
\method{print}{regimen}(x, ...)
}
\arguments{
\item{x}{regimen}

\item{...}{arguments to pass}
}
\value{
No return value, print function.
}
\description{
Print function for PKPDsim regimen
}
\keyword{internal}
