% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_json.R
\name{read_model_json}
\alias{read_model_json}
\title{Read model definition from JSON}
\usage{
read_model_json(path)
}
\arguments{
\item{path}{Path to JSON file}
}
\value{
List containing contents of original JSON file
}
\description{
Does some substitution of escaped characters in strings in the JSON file,
then converts to a list with \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}
}
