% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reduce_state_numbers}
\alias{reduce_state_numbers}
\title{R starts counting vector indices at 1, c++ starts at 0,
so reduce all state numbers in the Cpp function definition by 1}
\usage{
reduce_state_numbers(ode_def)
}
\description{
R starts counting vector indices at 1, c++ starts at 0,
so reduce all state numbers in the Cpp function definition by 1
}

