% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen_to_nm.R
\name{regimen_to_nm}
\alias{regimen_to_nm}
\title{Convert PKPDsim regimen to NONMEM table (doses only)}
\usage{
regimen_to_nm(
  reg = NULL,
  dose_cmt = 1,
  n_ind = 1,
  t_obs = NULL,
  obs_cmt = 1,
  bioav = NULL
)
}
\arguments{
\item{reg}{\code{PKPDsim} regimen, created using \code{new_regimen()} function}

\item{dose_cmt}{dosing compartment, if not specified in \code{reg} object. When
supplied explicitly, takes precedence over the compartments in \code{reg}.}

\item{n_ind}{repeat for \code{n_ind} subjects}

\item{t_obs}{add observation time(s)}

\item{obs_cmt}{observation compartment for added observation time(s)}

\item{bioav}{bioavailability (numeric vector, can not be a parameter)}
}
\value{
Data frame containing doses
}
\description{
Note: when bioavailability is used for insusions: NONMEM behaves differently from PKPDsim and Monolix,
in that rates are not automatically recomputed for infusions where bioavailabiliy also applies.
In PKPDsim/Monolix, for a bioavailability of 50\%, an AMT of 100 mg and rate of 100mg/hr would be recalculated
to 50 mg and 50 mg/hour, respectively, to keep the infusion length the same. This is not the case
in NONMEM. Therefore, the easiest way to work around this is for the user to compute the rate manually
in NONMEM using custom code. Therefore, we set the \code{RATE} column to \code{-1} for such infusions.
}
