% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparametrize.R
\name{reparametrize}
\alias{reparametrize}
\title{Reparametrize model parameters using a reparametrization defined within the model.}
\usage{
reparametrize(parameters, model)
}
\arguments{
\item{parameters}{list of model parameters}

\item{model}{PKPDsim model, compiled using `reparametrization` argument or in metadata object.}
}
\description{
Mostly useful for reparametrizing models into standard parametrizations, e.g. to
NONMEM TRANS or clinPK parametrizations.
}
