% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{search_replace_in_file}
\alias{search_replace_in_file}
\title{Find string and replace in file}
\usage{
search_replace_in_file(files = c(), find = NULL, replacement = NULL)
}
\arguments{
\item{files}{vector of files}

\item{find}{find what string, vector of character}

\item{replacement}{replace with what, vector of character, should be equal in length to `find`}
}
\value{
Function does not return a value but edits files on disk
}
\description{
Find string and replace in file
}
