% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_regimen.R
\name{shift_regimen}
\alias{shift_regimen}
\title{Remove n doses (from start) of PKPDsim regimen}
\usage{
shift_regimen(regimen, n = 1, reset_time = TRUE)
}
\arguments{
\item{regimen}{PKPDsim regimen created using `new_regimen()`}

\item{n}{number of doses to shift regimen}

\item{reset_time}{reset the remaining doses to start at t=0?}
}
\value{
Regimen with selected number of doses removed from start
}
\description{
Opposite of pop_regimen()
}
\seealso{
pop_regimen
}
