% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shift_state_indices.R
\name{shift_state_indices}
\alias{shift_state_indices}
\title{R starts counting vector indices at 1, c++ starts at 0,
so reduce all state numbers in the Cpp function definition by 1}
\usage{
shift_state_indices(ode_def, n = -1)
}
\arguments{
\item{ode_def}{ODE definition}

\item{n}{add/substract what number, default = -1}
}
\description{
R starts counting vector indices at 1, c++ starts at 0,
so reduce all state numbers in the Cpp function definition by 1
}

