% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ae.R
\name{sim_ae}
\alias{sim_ae}
\title{Simulate analytical PK equations}
\usage{
sim_ae(model = "pk_1cmt_iv", parameters, regimen, cov_model = NULL,
  covariates = NULL, t_obs = NULL)
}
\arguments{
\item{model}{model, choose from `pk_1cmt_iv`, `pk_1cmt_oral`, `pk_2cmt_iv`, `pk_2cmt_oral`.}

\item{parameters}{list of model parameters}

\item{regimen}{PKPDsim regimen}

\item{t_obs}{vector of observation times}
}
\description{
Simulate analytical PK equations
}
