% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_core_analytic.R
\name{sim_core_analytic}
\alias{sim_core_analytic}
\title{Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!
This is separate function for analytic equations only. See `sim_core()` for ODE function.`}
\usage{
sim_core_analytic(
  sim_object = NULL,
  model,
  duplicate_t_obs = FALSE,
  t_init = 0
)
}
\arguments{
\item{sim_object}{list with design and simulation parameters}

\item{model}{model object}

\item{duplicate_t_obs}{allow duplicate t_obs in output? E.g. for optimal design calculations when t_obs = c(0,1,2,2,3). Default is FALSE.}

\item{t_init}{time of initization of the ODE system. Usually 0.}
}
\description{
Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!
This is separate function for analytic equations only. See `sim_core()` for ODE function.`
}
