% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_list.R
\name{table_to_list}
\alias{table_to_list}
\title{Convert a table to a list}
\usage{
table_to_list(table)
}
\arguments{
\item{table}{data.frame}
}
\value{
List containing original table contents
}
\description{
Convert a table to a list
}
