% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Test a model}
\usage{
test_model(url, test_file, package, force = FALSE)
}
\arguments{
\item{url}{URL or file path to JSON representation of model}

\item{test_file}{Path to a .R file containing tests to run}

\item{package}{Package name}

\item{force}{Run tests even if model is not flagged for building? Defaults to
FALSE}
}
\value{
Runs test file for a model but does not return a value
}
\description{
Test a model
}
