% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pointer.R
\name{test_pointer}
\alias{test_pointer}
\title{Test if model still in memory}
\usage{
test_pointer(model)
}
\arguments{
\item{model}{pointer to model}
}
\value{
No return value
}
\description{
Test if model still in memory
}
