% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{ThreeCompIVinfusion}
\alias{ThreeCompIVinfusion}
\title{IV infusion- 3 compartment}
\usage{
ThreeCompIVinfusion(d)
}
\arguments{
\item{d}{data, Accepts a NONMEM style data frame for 1 subject with columns for TIME, AMT,MDV,RATE, RATEALL, DV, CL, V1, Q12, V2, Q13, V3}
}
\value{
Returns a dataframe with populated columns for A1, A2, A3,and DV
}
\description{
IV infusion- 3 compartment
}
