% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{ThreeCompIVinfusionMetab}
\alias{ThreeCompIVinfusionMetab}
\title{3-compartment IV infusion with first-order metabolite formation}
\usage{
ThreeCompIVinfusionMetab(d)
}
\arguments{
\item{d}{data}
}
\description{
3-compartment IV infusion with first-order metabolite formation
}
