% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{ThreeCompOralMetab}
\alias{ThreeCompOralMetab}
\title{first-order absorption- 3 compartment-Metabolite}
\usage{
ThreeCompOralMetab(d)
}
\arguments{
\item{d}{data}
}
\description{
first-order absorption- 3 compartment-Metabolite
}
