% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_ode.R
\name{translate_ode}
\alias{translate_ode}
\title{Translate a model from/to various PKPD simulators}
\usage{
translate_ode(code, auto = TRUE, from = NULL, to = NULL, verbose = TRUE)
}
\arguments{
\item{code}{character string with ODE code}

\item{auto}{is auto-detect syntax (`from`)}

\item{from}{from syntax}

\item{to}{to syntax}

\item{verbose}{verbose, `TRUE` or `FALSE`}
}
\value{
Translated PKPDsim or RxODE model
}
\description{
Currently only supports PKDPsim <--> RxODE
}
