% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/triangle_to_full.R
\name{triangle_to_full}
\alias{triangle_to_full}
\title{Convert triangle omega matrix to full omega matrix}
\usage{
triangle_to_full(vect)
}
\arguments{
\item{vect}{vector specifying triangle omega matrix}
}
\description{
Convert triangle omega matrix to full omega matrix
}

