% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{TwoCompIVinfusion}
\alias{TwoCompIVinfusion}
\title{IV infusion- 2 compartment}
\usage{
TwoCompIVinfusion(d)
}
\arguments{
\item{d}{data, accepts a NONMEM style data frame for 1 subject with columns for TIME, AMT,MDV, RATE, RATEALL, DV, CL, V1, Q, V2}
}
\value{
Returns a dataframe with populated columns for A1, A2, and DV
}
\description{
IV infusion- 2 compartment
}
