% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{TwoCompOral}
\alias{TwoCompOral}
\title{First-order absorption- 2 compartment}
\usage{
TwoCompOral(d)
}
\arguments{
\item{d}{data, accepts a NONMEM style data frame for 1 subject with columns for TIME, AMT,MDV,DV, CL, V2, Q, V3, KA & F1}
}
\value{
Returns a dataframe with populated columns for A1, A2, A3 and DV
}
\description{
First-order absorption- 2 compartment
}
\keyword{internal}
