% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ruv.R
\name{add_ruv}
\alias{add_ruv}
\title{Add residual variability to the dependent variable}
\usage{
add_ruv(x, ruv = list(), obs_type = 1)
}
\arguments{
\item{x}{dependent value without residual variability}

\item{ruv}{list specifying proportional, additive and/or exponential errors (\code{prop}, \code{add}, \code{exp})}

\item{obs_type}{vector of observation types}
}
\value{
Input vector with residual variability added
}
\description{
Add residual variability to the dependent variable
}
