% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adherence.R
\name{adherence_binomial}
\alias{adherence_binomial}
\title{Binomial adherence'}
\usage{
adherence_binomial(n = 100, p = 0.5)
}
\arguments{
\item{n}{number of days}

\item{p}{binomial probability}
}
\description{
Binomial adherence'
}

