% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adherence.R
\name{adherence_markov}
\alias{adherence_markov}
\title{Markov adherence'}
\usage{
adherence_markov(n = 100, p11 = 0.9, p01 = 0.7)
}
\arguments{
\item{n}{number of days}

\item{p11}{probability of staying adherent}

\item{p01}{probability of going from non-adherent to adherent state}
}
\description{
Markov adherence'
}

