% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan_process_infusion_doses.R
\name{advan_process_infusion_doses}
\alias{advan_process_infusion_doses}
\title{Add column RATEALL to ADVAN-style dataset to handle infusions}
\usage{
advan_process_infusion_doses(data)
}
\arguments{
\item{data}{ADVAN-style dataset, e.g. created using `advan_create_data`.}
}
\value{
Data frame containing additional RATEALL column.
}
\description{
Function adapted from code from Abuhelwa, Foster, Upton JPET 2015.
cleaned up and somewhat optimized. Can potentially be optimized more.
}
\references{
Abuhelwa, A. Y., Foster, D. J. R., Upton, R. N. (2015)
  ADVAN-style analytical solutions for common pharmacokinetic models. J
  Pharmacol Toxicol Methods 73:42-8. DOI: 10.1016/j.vascn.2015.03.004
}
