% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lagtime.R
\name{apply_lagtime}
\alias{apply_lagtime}
\title{Apply lagtime to a regimen}
\usage{
apply_lagtime(regimen, lagtime, parameters, cmt_mapping = NULL)
}
\arguments{
\item{regimen}{PKPDsim regimen}

\item{lagtime}{lagtime object, either single value / parameter name or vector of values/parameter names for all compartments.}

\item{parameters}{parameter list, required if parameters are specified.}

\item{cmt_mapping}{map of administration types to compartments, e.g. \code{list("oral" = 1, "infusion" = 2, "bolus" = 2)}.}
}
\value{
Original regimen with lagtime added to dose times
}
\description{
Apply lagtime to a regimen
}
