% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixed_parameters.R
\name{get_fixed_parameters}
\alias{get_fixed_parameters}
\title{Get fixed parameters}
\usage{
get_fixed_parameters(def)
}
\arguments{
\item{def}{Model definition as output by \code{\link[=read_model_json]{read_model_json()}}}
}
\description{
Get fixed parameters listed in model definition if present. If not present, use size of omega matrix to determine fixed parameters.
}
