% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_parameters_from_code.R
\name{get_parameters_from_code}
\alias{get_parameters_from_code}
\title{Get model parameters from code}
\usage{
get_parameters_from_code(code, declare_variables = NULL)
}
\arguments{
\item{code}{code}

\item{declare_variables}{declared variables}
}
\description{
Get model parameters from code
}

