% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_download}
\alias{github_download}
\title{Download a single file from GitHub, or get list of files}
\usage{
github_download(repo, path = "/README.md", branch = "master",
  auth_token = NULL, raw = FALSE, action = "download")
}
\arguments{
\item{repo}{repository, specified as `"owner/repo"`}

\item{path}{path to file}

\item{branch}{git branch}

\item{auth}{auth_token (for private repositories)}
}
\description{
Download a single file from GitHub, or get list of files
}
