% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_positive_definite.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Is matrix positive definite}
\usage{
is_positive_definite(x)
}
\arguments{
\item{x}{matrix, specified either as \code{vector} of lower triangle, or full matrix (as \code{matrix} class)}
}
\value{
TRUE if \code{x} is positive definite; FALSE otherwise.
}
\description{
Is matrix positive definite
}
