% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_from_api.R
\name{model_from_api}
\alias{model_from_api}
\title{Load model definition from API, and compile to R library}
\usage{
model_from_api(
  model = NULL,
  url = "http://localhost:8080/api",
  github = FALSE,
  repo = NULL,
  auth_token = NULL,
  verbose = TRUE,
  get_definition = FALSE,
  to_package = FALSE,
  force = FALSE,
  install_all = FALSE,
  run_tests = FALSE,
  ...
)
}
\arguments{
\item{model}{model id}

\item{url}{URL of API}

\item{github}{not a custom API but GitHub}

\item{repo}{GH repository specified as `owner/repo`}

\item{auth_token}{for private repositories}

\item{verbose}{verbosity (T/F)}

\item{get_definition}{return only the model definition, do not compile}

\item{to_package}{compile to package?}

\item{force}{force install even if same version number of model already installed.}

\item{install_all}{force install all, even if model inactive}

\item{run_tests}{run tests associated with model?}

\item{...}{arguments passed to `new_ode_model()` function}
}
\description{
Load model definition from API, and compile to R library
}
