% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_from_api.R
\name{model_from_api}
\alias{model_from_api}
\title{Load model definition from API, and compile to R library}
\usage{
model_from_api(
  url,
  model = NULL,
  nonmem = NULL,
  verbose = TRUE,
  get_definition = FALSE,
  to_package = FALSE,
  force = FALSE,
  install_all = FALSE,
  ...
)
}
\arguments{
\item{url}{URL or file path to JSON representation of model}

\item{model}{model id (used in messages)}

\item{nonmem}{URL or file path to NONMEM file}

\item{verbose}{verbosity (T/F)}

\item{get_definition}{return only the model definition, do not compile}

\item{to_package}{compile to package?}

\item{force}{force install even if same version number of model already installed.}

\item{install_all}{force install all, even if model inactive}

\item{...}{arguments passed to \code{new_ode_model()} function}
}
\value{
Model object created with \code{\link[=new_ode_model]{new_ode_model()}}
}
\description{
Load model definition from API, and compile to R library
}
