% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_locf.R
\name{na_locf}
\alias{na_locf}
\title{Fill in NAs with the previous non-missing value}
\usage{
na_locf(object, fromLast = FALSE)
}
\arguments{
\item{object}{an object}

\item{fromLast}{logical. Causes observations to be carried backward rather
than forward. Default is FALSE.}
}
\value{
Original object with NAs filled in
}
\description{
Inspired by zoo::na.locf0
}
