% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adherence.R
\name{new_adherence}
\alias{new_adherence}
\title{Adherence function}
\usage{
new_adherence(n = 100, type = "markov", markov = list(p11 = 0.75, p01 =
  0.75), p_binom = 0.7)
}
\arguments{
\item{n}{number of days to simulate}

\item{type}{type of adherence simulation, either "markov" or "binomial"}

\item{markov}{if markov-type, specify a list with markov transition probabilities}

\item{p_binom}{binomial probability of being adherent}
}
\description{
Adherence function
}

