% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_covariate_model.R
\name{new_covariate_model}
\alias{new_covariate_model}
\title{covariate model function}
\usage{
new_covariate_model(model = list())
}
\arguments{
\item{model}{covariate model specified as list}
}
\description{
covariate model function
}

