% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_regimen.R
\name{new_regimen}
\alias{new_regimen}
\title{Dose regimen for sim_ode}
\usage{
new_regimen(amt = 100, interval = 12, n = 3, times = NULL,
  type = "bolus", t_inf = NULL)
}
\arguments{
\item{amt}{dosing amount, either a single value (which will repeated for multiple doses), or a vector with doses for each administration}

\item{interval}{dosing interval (requires n as argument)}

\item{n}{number of doses (requires interval as argument)}

\item{times}{vector describing dosing times. Overrides specified times using interval and n arguments}

\item{type}{either "infusion" or "bolus" (default)}

\item{t_inf}{infusion time (if type==infusion)}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Create a dosing regimen for use with sim_ode
}
\examples{
r1 <- new_regimen(amt=50, interval=12, n=20) # dose 50mg, q12hrs for 10 days
r2 <- new_regimen(amt=50, times=c(0:19)*12)  # same, but using explicit times
r3 <- new_regimen(amt=c(rep(100,4), rep(50,16)), times=c(0:19)*12)  # first 4 doses higher dose
}
\seealso{
\link{sim_ode}
}

