% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_regimen_from_nm_dataset.R
\name{new_regimen_from_nm_dataset}
\alias{new_regimen_from_nm_dataset}
\title{Create a regimen from NM data}
\usage{
new_regimen_from_nm_dataset(data, reset_time = TRUE, first_only = FALSE)
}
\arguments{
\item{data}{NONMEM-type dataset}

\item{reset_time}{start time for each simulated patient at 0, irrespective of design in dataset}

\item{first_only}{use only design from first individual in dataset}
}
\description{
Create a regimen based on a NONMEM, or NONMEM-like dataset
}

