% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_obs_types.R
\name{parse_obs_types}
\alias{parse_obs_types}
\title{Parse observation types to simulation code}
\usage{
parse_obs_types(obs, initial = FALSE)
}
\arguments{
\item{obs}{specified observation object including at least a description of
which variable(s) are associated with a particular compartment, e.g.
`list(variable="CONC", scale="1")`.}

\item{initial}{is this for the initial code block in the C++ template that
initializes the variables and compartments (`TRUE`), or for the
second code block used for the rest of the dataset?}
}
\description{
Parse observation types to simulation code
}
\keyword{internal}
