% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_regimen.R
\name{parse_regimen}
\alias{parse_regimen}
\title{Parse regimen}
\usage{
parse_regimen(regimen, t_max, t_obs, t_tte, p, covariates)
}
\arguments{
\item{regimen}{regimen}

\item{t_max}{t_max}

\item{t_obs}{t_obs}

\item{t_tte}{t_tte}

\item{p}{parameters}

\item{covariates}{covariates}
}
\description{
Parse regimen
}

