% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkpd_1cmt_oral_viral.R
\name{pk_1cmt_oral_pd_viral}
\alias{pk_1cmt_oral_pd_viral}
\title{ODE system for PKPD - 1 compartment IV, HIV viral dynamics model (Perelson)}
\usage{
pk_1cmt_oral_pd_viral(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PKPD - 1 compartment IV, HIV viral dynamics model (Perelson)
}

