% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pk_tte_1cmt_oral_exp_hazard.R
\name{pk_tte_1cmt_oral_exp_hazard}
\alias{pk_tte_1cmt_oral_exp_hazard}
\title{ODE system for PK-TTE - 1 compartment oral + hazard}
\usage{
pk_tte_1cmt_oral_exp_hazard(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PK-TTE - 1 compartment oral + hazard
}

