% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkpd_hemtox.R
\name{pkpd_hemtox}
\alias{pkpd_hemtox}
\title{ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)}
\usage{
pkpd_hemtox(t, A, p)
}
\arguments{
\item{t}{time vector}

\item{A}{initial state vector}

\item{p}{parameters (list)}
}
\description{
ODE system for PKPD - 2 compartment IV, neutropenia model (Friberg et al. JCO 2002)
}

