% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PKPDsim-package.R
\docType{package}
\name{PKPDsim-package}
\alias{PKPDsim-package}
\title{PKPDsim package}
\description{
Simulate regimens for PKPD models defined by ODE systems
}
\author{
Ron Keizer \email{ronkeizer@gmail.com}
}

