% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_regimen.R
\name{pop_regimen}
\alias{pop_regimen}
\title{Remove n doses (from tail) of PKPDsim regimen}
\usage{
pop_regimen(regimen, n = 1)
}
\arguments{
\item{regimen}{PKPDsim regimen created using `new_regimen()`}

\item{n}{number of doses to pop from regimen}
}
\value{
Input regiment minus selected number of doses
}
\description{
Opposite of shift_regimen()
}
\seealso{
shift_regimen
}
