% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.PKPDsim.R
\name{print.PKPDsim}
\alias{print.PKPDsim}
\title{Print function for PKPDsim simulation function}
\usage{
\method{print}{PKPDsim}(x)
}
\arguments{
\item{x}{function}
}
\description{
Print function for PKPDsim simulation function
}

