% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_list}
\alias{print_list}
\title{Return a list in R syntax}
\usage{
print_list(x, wrapper = TRUE)
}
\arguments{
\item{x}{list to be printed}

\item{wrapper}{wrap in list object?}
}
\value{
Original list in R syntax
}
\description{
Return a list in R syntax
}
