% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_core.R
\name{sim_core}
\alias{sim_core}
\title{Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!}
\usage{
sim_core(sim_object = NULL, ode)
}
\arguments{
\item{sim_object}{list with design and simulation parameters}

\item{ode}{ode}
}
\description{
Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!
}

