% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_core.R
\name{sim_core}
\alias{sim_core}
\title{Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!}
\usage{
sim_core(sim_object = NULL, ode, duplicate_t_obs = FALSE, t_init = 0)
}
\arguments{
\item{sim_object}{list with design and simulation parameters}

\item{ode}{ode}

\item{duplicate_t_obs}{allow duplicate t_obs in output? E.g. for optimal design calculations when t_obs = c(0,1,2,2,3). Default is FALSE.}

\item{t_init}{time of initization of the ODE system. Usually 0.}
}
\value{
Data frame with simulation results
}
\description{
Only core function of the simulation function, always just returns observations.
Mostly useful for estimations / optimal design. Has no checks (for speed)!
}
