% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ode_shiny.R
\name{sim_ode_shiny}
\alias{sim_ode_shiny}
\title{Simulate ODE and create a Shiny app}
\usage{
sim_ode_shiny(...)
}
\arguments{
\item{...}{arguments passed to PKPDsimShiny::sim_ode_shiny()}
}
\description{
Simulate ODE and create a Shiny app
}
\seealso{
\link{sim_ode}
}

