% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_ode_shiny.R
\name{sim_ode_shiny}
\alias{sim_ode_shiny}
\title{Simulate ODE and create a Shiny app}
\usage{
sim_ode_shiny(name = "", ode = function() { }, parameters = list(),
  omega = NULL, omega_type = "exponential", n_ind = 1, regimen = NULL,
  A_init = NULL, step_size = 1, tmax = NULL,
  shiny_folder = "~/shiny-pkpd")
}
\arguments{
\item{ode}{function describing the ODE system}

\item{parameters}{}

\item{omega}{vector describing the lower-diagonal of the between-subject variability matrix}

\item{omega_type}{exponential or normal}

\item{n_ind}{number of individuals to simulate}

\item{regimen}{a regimen object created using the regimen() function}

\item{A_init}{vector with the initial state of the ODE system}

\item{step_size}{the step size between the observations (NOT the step size of the differential equation solver)}

\item{tmax}{maximum simulation time, if not specified will pick the end of the regimen as maximum}

\item{output}{vector specifying which compartment numbers to output}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Creates a VPC plot from observed and simulation data
}
\examples{
p <- list(CL = 38.48,
         V  = 7.4,
         Q2 = 7.844,
         V2 = 5.19,
         Q3 = 9.324,
         V3 = 111)

omega <- c(0.3,       # IIV CL
          0.1, 0.3)  # IIV V

sim_ode_shiny(ode = pk_3cmt_iv,
             par = p)
}
\seealso{
\link{sim_ode}
}

