% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ode_shiny.R
\name{sim_ode_shiny}
\alias{sim_ode_shiny}
\title{Simulate ODE and create a Shiny app}
\usage{
sim_ode_shiny(...)
}
\arguments{
\item{...}{arguments passed to PKPDsimShiny::sim_ode_shiny()}
}
\value{
No return value
}
\description{
This function has been deprecated and moved to a separate package at
https://github.com/ronkeizer/PKPDsimshiny.
}
\seealso{
\link{sim_ode}
}
