% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{ThreeCompIVinfusionMetab}
\alias{ThreeCompIVinfusionMetab}
\title{3-compartment IV infusion with first-order metabolite formation}
\usage{
ThreeCompIVinfusionMetab(d)
}
\arguments{
\item{d}{data, accepts a NONMEM style data frame for 1 subject with columns for TIME, AMT,MDV,RATE, RATEALL, DV, CL, V1, Q12, V2, Q13, V3, CLM,VM,km}
}
\value{
Returns a dataframe with populated columns for A1, A2, A3,and DV
}
\description{
3-compartment IV infusion with first-order metabolite formation
}
\keyword{internal}
