% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advan.R
\name{ThreeCompOralMetab}
\alias{ThreeCompOralMetab}
\title{first-order absorption- 3 compartment-Metabolite}
\usage{
ThreeCompOralMetab(d)
}
\arguments{
\item{d}{data, accepts a NONMEM style data frame for 1 subject with columns for TIME, AMT,MDV,DV, CL, V2, Q3, V3, Q4, V4, KA & F1}
}
\value{
Returns a dataframe with populated columns for A1, A2, A3, A4 and DV
}
\description{
first-order absorption- 3 compartment-Metabolite
}
\keyword{internal}
