% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_to_R_code.R
\name{vector_to_R_code}
\alias{vector_to_R_code}
\title{Transform a vector into a string that evaluates to the same vector}
\usage{
vector_to_R_code(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
character string of length 1
}
\description{
Collapses a vector into a comma-separated list with strings quoted
(and special characters escaped). A general purpose helper function
for writing new model code.
}
\keyword{internal}
